+++ desktop/cmp/tab/impl/Tab.ts
@@ -9,6 +9,7 @@ import {TabModel} from '@xh/hoist/cmp/tab';
 import {hoistCmp, refreshContextView, uses} from '@xh/hoist/core';
 import {elementFromContent} from '@xh/hoist/utils/react';
 import {useRef} from 'react';
+import {errorBoundary} from '@xh/hoist/cmp/error/ErrorBoundary';
 
 /**
  * Wrapper for contents to be shown within a TabContainer. This Component is used by TabContainer's
@@ -41,9 +42,12 @@ export const tab = hoistCmp.factory({
         return frame({
             display: isActive ? 'flex' : 'none',
             className,
-            item: refreshContextView({
-                model: refreshContextModel,
-                item: elementFromContent(content, {flex: 1})
+            item: errorBoundary({
+                model: model.errorBoundaryModel,
+                item: refreshContextView({
+                    model: refreshContextModel,
+                    item: elementFromContent(content, {flex: 1})
+                })
             })
         });
     }